requirements.txt
# Core
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
python-dotenv>=0.19.0
pydantic>=1.10.0
python-multipart>=0.0.5
httpx>=0.23.0

# Database
supabase>=2.0.0
python-jose[cryptography]>=3.3.0
psycopg2-binary>=2.9.3
sqlalchemy>=1.4.0
alembic>=1.7.7

# AI/ML
google-generativeai>=0.3.0
openai>=0.27.0
numpy>=1.21.0
pandas>=1.3.0

# Voice
livekit>=0.9.0
livekit-agents>=0.5.0
livekit-plugins-deepgram>=0.1.0

# Telephony
plivo>=4.32.0
twilio>=7.0.0

# Google Services
google-auth>=2.0.0
google-auth-oauthlib>=0.4.6
google-api-python-client>=2.0.0

# Utilities
python-dateutil>=2.8.2
pytz>=2021.3
phonenumbers>=8.12.0
python-slugify>=5.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-cov>=3.0.0

# Development
black>=22.0.0
isort>=5.0.0
mypy>=0.910.0
flake8>=4.0.0
